requirements.txt
paho-mqtt
pycryptodome
pandas
numpy
pyyaml
tqdm
matplotlib
seaborn